requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azure-functions-durable

numpy==1.20.1
pandas==1.1.5
azure-core==1.11.0
azure-storage-blob==12.8.0
azure-storage-file-datalake==12.3.0
pyodbc==4.0.30
pytest>=0.0.0
requests-futures==1.0.0
pymongo==3.11.2